Update colors on college template.
Change intro background color from the old gold color to the gray used in the 2015 templates.

Replace the intro header color with the new digital headline gold.

Change feature link colors to use the light gray border and black text used in the 2015 templates.

Replace the background of the single-tile header with the new gold and give it black text with a normal font-weight.

Give each tile link a black background color by default.