Fix keyboard navigation within the main navigation.
Make sub-dropdown menu in the top navigation accessible to screen readers. Links sub-dropdown menus are now always visible (display:block), but the links within them are visibly hidden.

When the sub-dropdown menu is hovered, the links become visible. If a a sub-dropdown menu link becomes focused, the individual link becomes visible.

Fix will need to be revisited in the future.